mindset-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
ruff==0.3.3
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-gcp-trace
python-json-logger
pytest
httpx<0.28.0